Sync Lyric: Refactor
